requirements.txt
# Dependências do projeto
streamlit
pandas
numpy
plotly
scikit-learn
requests
beautifulsoup4
lxml
tqdm
python-dateutil
psycopg2-binary  # adicionado
python-dotenv     # adicionado

